Initial commit: This is Synergy+MPX in what I belive to be a "working state".
Did not test it, but at the moment I don't think
this will work with Win/OSX/non-MPX versions of synergy.

Copy & Paste support does not work atm (this was out of 
scope of my original project).

There are still some FIXXMEs in the source code. Mostly
where I did a quick workaround to get going.



git-svn-id: http://synergy-plus.googlecode.com/svn@758 33c40d52-04ba-11de-a43f-ff96a2b7a9af